# add ParseDecimal()