I am very pleased with your package. The reason for this PR is to request an optimization for Parse() for short strings like "±nnnn.mmmm" (see benchmarks), which are very commonly used in fintech today. I did not find another way to contact you.