@woodsbury 2022-11-02
Hi, that's great to hear that you like the package. I can see the code you've provided is showing a nice performance improvement, and that the main reason for it seems to be that it's using an int64 for the operations instead of the 128-bit integer value the existing Parse() function uses.

I've taken a stab at introducing the same kind of optimisation to the existing parse functions with some changes on the branch feature/short-parse-optimise

Running your benchmarks seems to show with those changes I can get comparable performance to your ParseDecimal() function, and also a nice improvement even for numbers that are too big to fully parse into a 64-bit integer.

Before:
```
goos: windows
goarch: amd64
pkg: github.com/woodsbury/decimal128
cpu: Intel(R) Core(TM) i5-1035G7 CPU @ 1.20GHz
BenchmarkParse_short-8                   7151416               160.0 ns/op
BenchmarkParse_long-8                    6471175               201.7 ns/op
BenchmarkParseDecimal_short-8           21897409                54.43 ns/op
BenchmarkParseDecimal_long-8             6087670               216.2 ns/op
```

After:
```
goos: windows
goarch: amd64
pkg: github.com/woodsbury/decimal128
cpu: Intel(R) Core(TM) i5-1035G7 CPU @ 1.20GHz
BenchmarkParse_short-8                  33649734                34.99 ns/op
BenchmarkParse_long-8                   17107956                68.61 ns/op
BenchmarkParseDecimal_short-8           25255357                48.64 ns/op
BenchmarkParseDecimal_long-8            19025814                69.34 ns/op
```

I'd be keen for you to take a look and see if the changes on that branch meet your needs, and if so I can get it merged and a new version tagged.