@woodsbury 2022-11-03
That's awesome, I've merged that branch in and tagged a v0.4.5 version.